Refactor StarRating component to accept dynamic number of stars
The StarRating component has been refactored to accept a dynamic number of stars. This allows for greater flexibility in displaying star ratings based on different contexts. The "number" prop has been added to specify the total number of stars to render. This change improves the reusability and scalability of the StarRating component.